chore: move variant form helper
