+++ ui/src/types/Variant.ts
@@ -15,3 +15,14 @@ export interface IVariant extends IVariantBase {
 
 export type FilterableVariant = Pick<IVariant, 'id' | 'key' | 'name'> &
   IFilterable;
+
+export function toFilterableVariant(selected: IVariant | undefined) {
+  if (selected) {
+    return {
+      ...selected,
+      displayValue: selected.name,
+      filterValue: selected.id
+    };
+  }
+  return null;
+}
